Better styling

+++ src/app/pages/schedules/schedule-list/schedule-list.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit, Output } from '@angular/core';
+import { AlertService } from 'src/app/services/alert.service';
 import { ScheduleService } from 'src/app/services/schedule.service';
 import { ScheduleInterface } from 'src/app/shared/interfaces/schedule.interface';
 
@@ -15,7 +16,10 @@ export class ScheduleListComponent implements OnInit {
   idToEdit: number = 0;
   showScheduleForm: boolean = false;
 
-  constructor(private scheduleService: ScheduleService) { }
+  constructor(
+    private scheduleService: ScheduleService,
+    private alertService: AlertService
+  ) { }
 
   ngOnInit(): void {
     this.getSchdules();
@@ -26,10 +30,6 @@ export class ScheduleListComponent implements OnInit {
     this.scheduleService.list().subscribe(
       (success) => {
         this.schedules = success;
-        console.log(success);
-      },
-      (error) => {
-        alert('Erro ao recuperar lista de coisas legais')
       }
     )
   }
@@ -46,14 +46,28 @@ export class ScheduleListComponent implements OnInit {
     this.showScheduleForm = true;
   }
 
-  removeSchedule(id: number) {
-    this.scheduleService.delete(id).subscribe(
+  async removeSchedule(id: number) {
+    (await this.alertService.confirm('Você tem certeza?', 'Você não poderá desfazer essa ação', 'DELETAR')).subscribe(
       (success) => {
-        alert(success.message);
-      },
-      (error) => {
-        alert(error.message)
+        if (success.isConfirmed) {
+          this.scheduleService.delete(id).subscribe(
+            (success) => {
+              this.alertService.success('Sucesso', 'Agendamento excluído com sucesso', 'FECHAR');
+              this.getSchdules();
+            },
+            (error) => {
+              this.alertService.error('Erro', 'Não foi possível excluir o agendamento', 'FECHAR');
+            }
+          );
+        }
       }
     )
+  }
+
+  eventForm(event: any) {
+    if (event === 'update') {
+      this.getSchdules();
+    }
+    this.showScheduleForm = false;
   }
 }
